issue #162 refactor: directory change
