+++ src/components/Setting/WithdrawalTab/WithdrawalTab.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import BaseModal from "@/components/Base/BaseModal/BaseModal";
+import BaseModal from "@/components/Common/BaseModal/BaseModal";
 import { withdrawMembership } from "@/features/auth/auth-service";
 import { closeModal, openModal } from "@/features/ui/ui-slice";
